Moved the convertToCSV fn to support/utils.js
